+++ src/main/java/by/tms/springstore/repository/ProductRepository.java
@@ -4,15 +4,16 @@ import by.tms.springstore.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface ProductRepository extends JpaRepository<Product, Integer> {
 
     List<Product> findAll();
     //        List<Product> getProducts ();
 
-    Product findById(int id);
+    Product findById(Long id);
     //    Product getProductById(int id);
 
-    List<Product> findAllByCategoryId(int categoryId);
+    List<Product> findAllByCategoryId(Long categoryId);
     //    List<Product> getProductsByCategoryId(int categoryId);
 }
