+++ src/main/java/by/tms/springstore/service/CategoryService.java
@@ -3,12 +3,13 @@ package by.tms.springstore.service;
 import by.tms.springstore.domain.Category;
 
 import java.util.List;
+import java.util.Optional;
 
 public interface CategoryService {
 
     List<Category> getCategories();
 
-    Category getCategoryById(int id);
+    Category getCategoryById(Long id);
 
     Category getCategoryByName(String name);
 }
